+++ soya-starter/src/components/common/PhotoCaption/PhotoCaption.js
@@ -3,7 +3,7 @@ import {prop} from 'soya/lib/helper';
 
 // We can use webpack's localized CSS to ensure there is no naming clash
 // between our components.
-var style = prop(require('./PhotoCaption.css'), 'locals');
+var style = require('./PhotoCaption.css');
 
 export default class PhotoCaption extends React.Component {
   render() {